combine appconfig tests
